changing int to unsigned long for the n_col

+++ t-sne.h
@@ -13,8 +13,8 @@ extern "C" {
 
 double *ts_fit(int N, int n_in, double *x, int n_out, double theta, double perplexity, int seed);
 
-double *sann_data_read_1d(const char *fn, int *n, int *n_col, char ***row_names, char ***col_names);
-float **sann_data_read(const char *fn, int *n_, int *n_col_, char ***row_names, char ***col_names);
+double *sann_data_read_1d(const char *fn, int *n, unsigned long *n_col, char ***row_names, char ***col_names);
+float **sann_data_read(const char *fn, int *n_, unsigned long *n_col_, char ***row_names, char ***col_names);
 void sann_free_names(int n, char **s);
 void sann_free_vectors(int n, float **x);
